moved websockets code to httpsConnection.js
